% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcGPPGUI.R
\name{CalcGPPGUI}
\alias{CalcGPPGUI}
\title{Function 'CalcGPPGUI' calls GUI to calculate and plot GPP data based on the function 'CalcGPP'.
It calls the GUI that will ask user to upload the data to be used in 'CalcGPP' function
(see 'CalcGPP' function for further details).
GUI window is created using 'guiv' function from  'fgui' package. Please ensure to cite the required packages.}
\usage{
CalcGPPGUI(CalcGPP)
}
\description{
Function 'CalcGPPGUI' calls GUI to calculate and plot GPP data based on the function 'CalcGPP'.
It calls the GUI that will ask user to upload the data to be used in 'CalcGPP' function
(see 'CalcGPP' function for further details).
GUI window is created using 'guiv' function from  'fgui' package. Please ensure to cite the required packages.
}
