% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcRecoMOD.R
\name{CalcRecoMOD}
\alias{CalcRecoMOD}
\title{Function 'CalcRecoMOD' calculates the modelled Reco using optimised R10 values by applying the
modified Arrhenius equation adapted from Lloyd and Taylor (1994).
It will write the data into new csv file ('RecoMOD.csv').
At the end it calls next GUI that will ask user to upload the RecoMOD data and will plot the data.}
\usage{
CalcRecoMOD(csvFilename_initialData, R10o)
}
\description{
References:
\itemize{
\item Lloyd, J., Taylor, J.A. (1994) On the temperature dependence of soil respiration. Functional Ecology, 8, 315-323. DOI: https://doi.org/10.2307/2389824".
}
}
\details{
GUI window is created using 'guiv' function from  'fgui' package. Please ensure to cite the required packages.
}
