% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiCallb_PartECd.R
\name{guiCallb_PartECd}
\alias{guiCallb_PartECd}
\title{The 'guiCallb_PartECd' is the GUI callback function required for creating the GUI
that supports 'PartECdata' function (see also the 'PartECdata' function).
It requires 'fgui' package. Please ensure to cite the required packages.}
\usage{
guiCallb_PartECd(arg)
}
\description{
The 'guiCallb_PartECd' is the GUI callback function required for creating the GUI
that supports 'PartECdata' function (see also the 'PartECdata' function).
It requires 'fgui' package. Please ensure to cite the required packages.
}
