% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiCallb_RecoM.R
\name{guiCallb_RecoM}
\alias{guiCallb_RecoM}
\title{The 'guiCallb_RecoM' is the GUI callback function required for creating the GUI
that supports 'CalcRecoMOD' function (see also the 'CalcRecoMOD' function).
It requires 'fgui' package. Please ensure to cite the required packages.}
\usage{
guiCallb_RecoM(arg)
}
\description{
The 'guiCallb_RecoM' is the GUI callback function required for creating the GUI
that supports 'CalcRecoMOD' function (see also the 'CalcRecoMOD' function).
It requires 'fgui' package. Please ensure to cite the required packages.
}
