% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiCallbackgpp.R
\name{guiCallbackgpp}
\alias{guiCallbackgpp}
\title{The 'guiCallbackgpp' is the GUI callback function required for creating the GUI
that supports 'OptimLight' function.
It requires 'fgui' package. Please ensure to cite the required packages.}
\usage{
guiCallbackgpp(arg)
}
\description{
The 'guiCallbackgpp' is the GUI callback function required for creating the GUI
that supports 'OptimLight' function.
It requires 'fgui' package. Please ensure to cite the required packages.
}
