% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiCallbackreco.R
\name{guiCallbackreco}
\alias{guiCallbackreco}
\title{The 'guiCallbackreco' is the GUI callback function required for creating the GUI
that supports 'ProcessDataReco' function.
It requires 'fgui' package. Please ensure to cite the required packages.}
\usage{
guiCallbackreco(arg)
}
\description{
The 'guiCallbackreco' is the GUI callback function required for creating the GUI
that supports 'ProcessDataReco' function.
It requires 'fgui' package. Please ensure to cite the required packages.
}
