% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiCallbackRecoMOD.R
\name{guiCallbackRecoMOD}
\alias{guiCallbackRecoMOD}
\title{The 'guiCallbackRecoMOD' is the GUI callback function required for creating the GUI
that supports 'Plot_RecoMod_T' function for plotting RecoMOD.
It requires 'fgui' package. Please ensure to cite the required packages.}
\usage{
guiCallbackRecoMOD(arg)
}
\description{
The 'guiCallbackRecoMOD' is the GUI callback function required for creating the GUI
that supports 'Plot_RecoMod_T' function for plotting RecoMOD.
It requires 'fgui' package. Please ensure to cite the required packages.
}
