% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModGFillGUI.R
\name{ModGFillGUI}
\alias{ModGFillGUI}
\title{Function 'ModGFillGUI' calls GUI to model GPP using obtained optimised parameter values (α, Amax, γ)
used in the light-response function adapted from Rabinowitch (1951), reported in Gilmanov et al. (2003).
The GUI will call function 'ModGPP' (see 'ModGPP'function for further details).}
\usage{
ModGFillGUI(ModGPP)
}
\description{
References:
\itemize{
\item Rabinowitch, E.I. (1951) Photosynthesis and Related Processes. Interscience Publishers.
\item Gilmanov, T.G., Verma, S.B., Sims, P.L., Meyers, T.P., Bradford, J.A., Burba, G.G., Suyker, A.E. (2003) Gross primary production and light response parameters of four Southern Plains ecosystems estimated using long-term CO2-flux tower measurements. Global Biogeochemical Cycles, 17/2, 1071, DOI:10.1029/2002GB002023.
}
}
\details{
GUI window is created using 'guiv' function from  'fgui' package. Please ensure to cite the required packages.
}
