% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimLight.R
\name{OptimLight}
\alias{OptimLight}
\title{Function 'OptimLight' is used to optimise parameters (α, Amax, γ) from light-response GPP function,
which is adapted from non-rectangular hyperbola (NRH) model by Rabinowitch (1951),
reported in Gilmanov et al. (2003); (for details on adapted NRH model equation see function 'ModGPP').}
\usage{
OptimLight(csvFilename_gpp, α0, Amax0, γ0)
}
\description{
Function 'FunGPPMod'  represents the light-response function equation (i.e. the adapted NRH model by
Rabinowitch (1951), reported in Gilmanov et al. (2003)).
The GPP is pre-modelled using function 'FunGPPMod' and initial parameter (α0, Amax0, γ0) values.
The  initial parameter values (α0 = 0.08, Amax = 15, γ= 0.5) are manually entered via
GUI that is called in function 'CalcGPP' (see also function 'CalcGPP' for further details - this function
at the end calls GUI that will call function 'OptimLight).
}
\details{
x0 represents initial values -  the parameter values (α, Amax, γ), which are
at the start  filled with values  #' x= c(0.08, 15, 0.5) prior to optimisation and the
initial variable values are:  α = 0.08 , Amax =  15, γ = 0.5'
Next, the parameter values (α, Amax, γ) are optimised (using function 'optim'
from package 'stats' by minimising the residual sum of squares from computed and pre-modelled GPP.
The optimisation (using function 'optim') is performed by applying the 'BFGS' method/algorithm
(Broyden, 1970; Fletcher, 1970; Goldfarb, 1970; Shanno, 1970).

References:
\itemize{
\item Rabinowitch, E.I. (1951) Photosynthesis and Related Processes. Interscience Publishers.
\item Gilmanov, T.G., Verma, S.B., Sims, P.L., Meyers, T.P., Bradford, J.A., Burba, G.G., Suyker, A.E. (2003) Gross primary production and light response parameters of four Southern Plains ecosystems estimated using long-term CO2-flux tower measurements. Global Biogeochemical Cycles, 17/2, 1071, DOI:10.1029/2002GB002023.
\item Broyden, C.G. (1970) The convergence of a class of double rank minimization algorithms: 2. The new algorithm. J. Inst. Math. Appl. 6, 222–231.
\item Fletcher, R. (1970) A new approach to variable metric algorithms. Comput. J. 13, 317–322.
\item Goldfarb, D. (1970) A family of variable metric methods derived by variational means. Math. Comp. 24, 23–26.
\item Shanno, D.F. (1970) Conditioning of quasi-Newton methods for function minimization. Math. Comp. 24, 647–650.
}

Please ensure to also cite the required packages.
}
