% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimR10.R
\name{OptimR10}
\alias{OptimR10}
\title{Function 'OptimR10' reads the nighttime flux data ('Reco_night.csv'),
extracts variables T and Reco, defines X0 (from GUI input), which is the chosen value  that
represents initial arbitrary value for R10 prior to optimisation.}
\usage{
OptimR10(csvFilename_reco, x0)
}
\description{
Function 'FunReco_Mod' is the modified Arrhenius equation adapted from Lloyd and Taylor (1994),
which is used to calculate modelled Reco:
Reco= R10  ·   e ^ (E0  ·  ( 1/(283.2 -To) – 1/(T-To))) ,
where: Reco is ecosystem respiration, R10 is respiration rate at 10°C  (µmol CO2 m^(-2) s^(-1)),
^ is power-sign; E0 is activation energy (309 K), temperature T0 = 230 K,
and T is  measured temperature in K.
Optimisation  of R10 is performed using function 'optim' from R 'stats' package
by applying the 'L-BFGS-B' method/algorithm of Byrd et al. (1995).
The output (optimised R10) is  written in the R-Console.
}
\details{
References:
\itemize{
\item Lloyd, J., Taylor, J.A. (1994) On the temperature dependence of soil respiration. Functional Ecology, 8, 315-323. DOI: https://doi.org/10.2307/2389824".
\item Byrd, R.H., Lu, P., Nocedal, J., Zhu, C. (1995) A Limited Memory Algorithm for Bound Constrained Optimization.  16, 1190-1208.
}

Please ensure to also cite the required packages.
}
