% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_gapfilledNEE.R
\name{Plot_gapfilledNEE}
\alias{Plot_gapfilledNEE}
\title{Function 'Plot_gapfilledNEE'  plots the gap-filled NEE (blue colour) together
with observed/measured NEE that contains gaps (green colour); (see explanation in function 'GFillData').}
\usage{
Plot_gapfilledNEE(csvFileGapFilled)
}
\description{
Please ensure to cite the required packages.
}
