% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessDataReco.R
\name{ProcessDataReco}
\alias{ProcessDataReco}
\title{Function 'ProcessDataReco' reads the nighttime flux data ('Reco_night.csv'),
extracts variables T and Reco and plots them (see explanation in function 'PartECdata').
Please ensure to cite the required packages.}
\usage{
ProcessDataReco(csvFilename_reco)
}
\description{
Function 'ProcessDataReco' reads the nighttime flux data ('Reco_night.csv'),
extracts variables T and Reco and plots them (see explanation in function 'PartECdata').
Please ensure to cite the required packages.
}
